add some functionalities in  profile
