+++ src/components/UI/Navbar.js
@@ -5,13 +5,15 @@ import Button from "./Button";
 import InputField from "./FormField";
 import Logo from "../../assets/images/Logo.png";
 import { AiOutlineSearch } from "react-icons/ai";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { Avatar } from "./Avatar";
 import { useEffect, useState } from "react";
+import { resetuser } from "../../features/user/userSlice";
 
 export default function Navbar() {
   const location = useLocation();
   const history = useHistory();
+  const dispatch = useDispatch();
   const current_user = useSelector(state => state.session.user)
   const avatar = (<div css={css`
                             display: flex;
@@ -26,6 +28,7 @@ export default function Navbar() {
                             }
                           `} onClick={()=>
                           {
+                            dispatch(resetuser(current_user.id))
                             history.push(`/profile`)
                           }}>
                    <Avatar src={current_user.avatar_url} type="small"></Avatar>
